Fetch and display areas from backend in AreasList page
