+++ Frontend/src/App.jsx
@@ -6,8 +6,8 @@ import {Router} from "./Router.jsx";
 import {NavBarMobile} from "./NavBarMobile.jsx";
 import {NavBarDesktop} from "./NavBarDesktop.jsx";
 import {IconCalendarEvent, IconHeart, IconHome, IconMap} from "@tabler/icons-react";
-import { useState, useEffect } from 'react';
-import axios from 'axios';
+import { AreasProvider } from './context/AreasProvider';
+
 
 const links = [
   {path: '/', label: 'Home', icon: IconHome},
@@ -20,23 +20,10 @@ const App = () => {
   const [opened, {toggle}] = useDisclosure();
   const navigate = useNavigate();
 
-  const [areas, setAreas] = useState([]);
 
-  useEffect(() => {
-    const fetchAreas = async () => {
-      try {
-        const response = await axios.get('/areas');
-        setAreas(response.data);
-        console.log('Fetched areas:', response.data); // Optional debug
-      } catch (error) {
-        console.error('Error fetching areas:', error);
-      }
-    };
-
-    fetchAreas();
-  }, []);
 
   return (
+    <AreasProvider>
     <AppShell
       header={{height: 60}}
       navbar={{width: 300, breakpoint: 'sm', collapsed: {desktop: true, mobile: !opened}}}
@@ -70,6 +57,7 @@ const App = () => {
         </div>
       </AppShell.Footer>
     </AppShell>
+    </AreasProvider>
   );
 }
