+++ Frontend/src/context/AreasContext.jsx
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+
+// Create the context with an initial empty array (we will provide data here later)
+export const AreasContext = createContext([]);
